Added a shortcut for no-wildcard globs in HDFS
