+++ src/Common/parseGlobs.cpp
@@ -17,6 +17,11 @@
 
 namespace DB
 {
+namespace ErrorCodes
+{
+    extern const int BAD_ARGUMENTS;
+}
+
 /* Transforms string from grep-wildcard-syntax ("{N..M}", "{a,b,c}" as in remote table function and "*", "?") to perl-regexp for using re2 library for matching
  * with such steps:
  * 1) search intervals like {0..9} and enums like {abc,xyz,qwe} in {}, replace them by regexp with pipe (expr1|expr2|expr3),
@@ -115,5 +120,68 @@ std::string makeRegexpPatternFromGlobs(const std::string & initial_str_with_glob
         previous = letter;
     }
     return buf_final_processing.str();
+}
+
+void expandSelector(const std::string & path, Strings & for_match_paths_expanded)
+{
+    /// regexp for {expr1,expr2,expr3}, expr.. should be without "{", "}", "*" and ","
+    static const re2::RE2 selector_regex(R"({([^{}*,]+,[^{}*]*[^{}*,])})");
+
+    std::string_view path_view(path);
+    std::string_view matched;
+
+    if (RE2::FindAndConsume(&path_view, selector_regex, &matched))
+        std::string buffer(matched);
+    else
+    {
+        for_match_paths_expanded.push_back(path);
+        return;
+    }
+
+    Strings expanded_paths;
+
+    std::vector<size_t> anchor_positions = {};
+    bool opened = false, closed = false;
+
+    for (std::string::const_iterator it = path.begin(); it != path.end(); it++)
+    {
+        if (*it == '{')
+        {
+            if (opened)
+                throw Exception(ErrorCodes::BAD_ARGUMENTS,
+                                "Unexpected '{{' found in path '{}' at position {}.", path, std::distance(path.begin(), it));
+            anchor_positions.push_back(std::distance(path.begin(), it));
+            opened = true;
+        }
+        else if (*it == '}')
+        {
+            if (!opened)
+                throw Exception(ErrorCodes::BAD_ARGUMENTS,
+                                "Unexpected '}}' found in path '{}' at position {}.", path, std::distance(path.begin(), it));
+            anchor_positions.push_back(std::distance(path.begin(), it));
+            closed = true;
+            break;
+        }
+        else if (*it == ',')
+        {
+            if (!opened)
+                throw Exception(ErrorCodes::BAD_ARGUMENTS,
+                                "Unexpected ',' found in path '{}' at position {}.", path, std::distance(path.begin(), it));
+            anchor_positions.push_back(std::distance(path.begin(), it));
+        }
+    }
+    if (!opened || !closed)
+        throw Exception(ErrorCodes::BAD_ARGUMENTS,
+                        "Invalid {{}} glob in path {}.", path);
+
+    std::string common_prefix = path.substr(0, anchor_positions[0]);
+    std::string common_suffix = path.substr(anchor_positions[anchor_positions.size()-1] + 1);
+    for (size_t i = 1; i < anchor_positions.size(); ++i)
+    {
+        std::string expanded_matcher = common_prefix
+                                       + path.substr(anchor_positions[i-1] + 1, (anchor_positions[i] - anchor_positions[i-1] - 1))
+                                       + common_suffix;
+        expandSelector(expanded_matcher, for_match_paths_expanded);
+    }
 }
 }
